chore(trace): move tracer module
