+++ ddtrace/opentracer/span.py
@@ -22,7 +22,7 @@ from .tags import Tags
 
 
 if TYPE_CHECKING:  # pragma: no cover
-    from .tracer import Tracer  # noqa:F401
+    from ddtrace._trace.tracer import Tracer  # noqa:F401
 
 
 _TagNameType = Union[Text, bytes]
